+++ workschd-api/persist/src/main/java/com/voyagerss/persist/dto/UnavailableDatesDTO.java
@@ -12,7 +12,7 @@ public class UnavailableDatesDTO implements Serializable {
     private static final long serialVersionUID = 1L;
     private Long id;
 
-    private Integer accountId;
+    private Long accountId;
 
     private LocalDate date;
